Move as_role to utils.
