+++ integration_tests/index/test_common_psql.py
@@ -6,13 +6,17 @@
 import pytest
 from sqlalchemy import text
 
-
 # These tests use an empty/uninitialised database.  Doesn't matter whether it is postgis or postgres.
+
 
 @pytest.fixture
 def specific_user(uninitialised_postgres_db, cfg_env):
-    from datacube.drivers.common_psql import create_schema, transfers_required, as_role, ensure_role, create_user, drop_users
-    from datacube.drivers.postgis._core import UserRole, SCHEMA_NAME
+    from datacube.drivers.common_psql import (
+        create_user,
+        drop_users,
+        ensure_role,
+    )
+    from datacube.drivers.postgis._core import UserRole
 
     engine = uninitialised_postgres_db._engine
     with engine.connect() as conn:
@@ -39,25 +43,41 @@ def bare_user(uninitialised_postgres_db, cfg_env):
 
 @pytest.mark.parametrize("datacube_env_name", ("postgis", "postgis3"))
 def test_transfer_perms(uninitialised_postgres_db, specific_user):
-    from datacube.drivers.common_psql import create_schema, drop_schema, transfers_required, as_role, transfer_ownership
+    from datacube.drivers.common_psql import (
+        as_role,
+        create_schema,
+        drop_schema,
+        transfer_ownership,
+        transfers_required,
+    )
     from datacube.drivers.postgis._core import SCHEMA_NAME
 
     engine = uninitialised_postgres_db._engine
     with engine.connect() as conn:
         create_schema(conn, SCHEMA_NAME)
-        conn.execute(text(f"grant all privileges on schema {SCHEMA_NAME} to {specific_user}"))
+        conn.execute(
+            text(f"grant all privileges on schema {SCHEMA_NAME} to {specific_user}")
+        )
         conn.execute(text(f"grant create on schema {SCHEMA_NAME} to {specific_user}"))
         conn.commit()
         with as_role(conn, specific_user) as conn:
-            conn.execute(text(f"create table {SCHEMA_NAME}.test_table (id serial primary key)"))
+            conn.execute(
+                text(f"create table {SCHEMA_NAME}.test_table (id serial primary key)")
+            )
 
         try:
-            transfers = transfers_required(conn, 'odc_admin', SCHEMA_NAME, "tables", prefix="test_")
+            transfers = transfers_required(
+                conn, "odc_admin", SCHEMA_NAME, "tables", prefix="test_"
+            )
             assert transfers == [("test_table", specific_user)]
 
-            transfer_ownership(conn, SCHEMA_NAME, "test_table", specific_user, "odc_admin", "tables")
+            transfer_ownership(
+                conn, SCHEMA_NAME, "test_table", specific_user, "odc_admin", "tables"
+            )
 
-            transfers = transfers_required(conn, 'odc_admin', SCHEMA_NAME, "tables", prefix="test_")
+            transfers = transfers_required(
+                conn, "odc_admin", SCHEMA_NAME, "tables", prefix="test_"
+            )
             assert transfers == []
         finally:
             drop_schema(conn, SCHEMA_NAME)
@@ -66,6 +86,7 @@ def test_transfer_perms(uninitialised_postgres_db, specific_user):
 @pytest.mark.parametrize("datacube_env_name", ("postgis", "postgis3"))
 def test_has_roles(uninitialised_postgres_db, specific_user):
     from datacube.drivers.common_psql import has_roles
+
     engine = uninitialised_postgres_db._engine
     with engine.connect() as conn:
         assert has_roles(conn, [specific_user, "odc_admin", "odc_manage", "odc_user"])
@@ -75,16 +96,17 @@ def test_has_roles(uninitialised_postgres_db, specific_user):
 def test_ensure_role(uninitialised_postgres_db, bare_user):
     from datacube.drivers.common_psql import as_role, ensure_role, has_role
     from datacube.drivers.postgis._core import UserRole
+
     engine = uninitialised_postgres_db._engine
     with engine.connect() as conn:
         assert has_role(conn, "odc_admin", with_create_role=True)
-        conn.execute(text(f"alter role odc_admin with nocreaterole"))
+        conn.execute(text("alter role odc_admin with nocreaterole"))
         assert not has_role(conn, "odc_admin", with_create_role=True)
         ensure_role(conn, UserRole.ADMIN)
         assert has_role(conn, "odc_admin", with_create_role=True)
 
     with engine.connect() as conn:
-        conn.execute(text(f"alter role odc_admin with nocreaterole"))
+        conn.execute(text("alter role odc_admin with nocreaterole"))
         assert not has_role(conn, "odc_admin", with_create_role=True)
         with as_role(conn, bare_user) as conn:
             assert not ensure_role(conn, UserRole.ADMIN)
@@ -93,7 +115,7 @@ def test_ensure_role(uninitialised_postgres_db, bare_user):
 
 @pytest.mark.parametrize("datacube_env_name", ("postgis", "postgis3"))
 def test_create_user(uninitialised_postgres_db, specific_user, bare_user):
-    from datacube.drivers.common_psql import create_user, as_role
+    from datacube.drivers.common_psql import as_role, create_user
     from datacube.drivers.postgis._core import UserRole
 
     engine = uninitialised_postgres_db._engine
@@ -106,12 +128,11 @@ def test_create_user(uninitialised_postgres_db, specific_user, bare_user):
 
 @pytest.mark.parametrize("datacube_env_name", ("postgis", "postgis3"))
 def test_drop_users(uninitialised_postgres_db, specific_user, bare_user):
-    from datacube.drivers.common_psql import drop_users, as_role
+    from datacube.drivers.common_psql import as_role, drop_users
 
     engine = uninitialised_postgres_db._engine
-    with engine.connect() as conn:
-        with as_role(conn, bare_user) as conn:
-            assert not drop_users(conn, [specific_user])
+    with engine.connect() as conn, as_role(conn, bare_user) as conn:
+        assert not drop_users(conn, [specific_user])
 
 
 @pytest.mark.parametrize("datacube_env_name", ("postgis", "postgis3"))
@@ -122,10 +143,24 @@ def test_as_role(uninitialised_postgres_db, specific_user, bare_user):
     with engine.connect() as conn:
         db_user = conn.execute(text("select quote_ident(current_user)")).scalar()
         with as_role(conn, None) as conn:
-            assert conn.execute(text("select quote_ident(current_user)")).scalar() == db_user
+            assert (
+                conn.execute(text("select quote_ident(current_user)")).scalar()
+                == db_user
+            )
         with as_role(conn, specific_user) as conn:
-            assert conn.execute(text("select quote_ident(current_user)")).scalar() == specific_user
+            assert (
+                conn.execute(text("select quote_ident(current_user)")).scalar()
+                == specific_user
+            )
             with as_role(conn, bare_user) as conn:
-               assert conn.execute(text("select quote_ident(current_user)")).scalar() == bare_user
-            assert conn.execute(text("select quote_ident(current_user)")).scalar() == specific_user
-        assert conn.execute(text("select quote_ident(current_user)")).scalar() == db_user
+                assert (
+                    conn.execute(text("select quote_ident(current_user)")).scalar()
+                    == bare_user
+                )
+            assert (
+                conn.execute(text("select quote_ident(current_user)")).scalar()
+                == specific_user
+            )
+        assert (
+            conn.execute(text("select quote_ident(current_user)")).scalar() == db_user
+        )
